CompareIT-api/requirements.txt
Django==1.5.4
MySQL-python==1.2.4
PIL==1.1.7
South==0.8.2
argparse==1.2.1
astroid==1.0.0
coverage==3.6
distribute==0.6.34
django-jenkins==0.14.1
djangorestframework==2.3.8
docutils==0.11
logilab-common==0.60.0
pep8==1.4.6
pyflakes==0.7.3
pylint==1.0.0
wsgiref==0.1.2
